# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handler.todoCreateInput:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  handler.todoOutput:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.todoUpdateInput:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: API for managing todo items
  title: Todo API
  version: "1.0"
paths:
  /todos:
    get:
      description: Retrieve all todo items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.todoOutput'
            type: array
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Create a new todo item
      parameters:
      - description: Todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/handler.todoCreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.todoOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a todo
      tags:
      - todos
  /todos/{id}:
    delete:
      description: Delete a todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete a todo by ID
      tags:
      - todos
    get:
      description: Retrieve a todo item by its ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.todoOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Update an existing todo item
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated todo data
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/handler.todoUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.todoOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update a todo
      tags:
      - todos
swagger: "2.0"