# internal/infra/handler/todo_get_by_id.go
package handler

import (
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/wellingtonlope/todo-api/internal/app/usecase/todo"
)

type (
	TodoGetByID struct {
		getByID todo.GetByID
	}
)

func NewTodoGetByID(getByID todo.GetByID) *TodoGetByID {
	return &TodoGetByID{getByID: getByID}
}

// @Summary Get a todo by ID
// @Description Retrieve a todo item by its ID
// @Tags todos
// @Produce json
// @Param id path string true "Todo ID"
// @Success 200 {object} todoOutput
// @Failure 404 {object} ErrorResponse
// @Router /todos/{id} [get]
func (h *TodoGetByID) Handle(c echo.Context) error {
	id := c.Param("id")
	output, err := h.getByID.Handle(c.Request().Context(), id)
	if err != nil {
		return err
	}
	return c.JSON(http.StatusOK, todoOutputFromUsecase(output))
}

func (h *TodoGetByID) Path() string {
	return "/todos/:id"
}

func (h *TodoGetByID) Method() string {
	return http.MethodGet
}